Changes the package target folder
