--- capstone/src/main/java/com/google/univiz/collegescorecarddata/CollegeScoreCard.java
@@ -1,19 +1,0 @@
-package com.google.univiz.collegescorecarddata;
-
-import java.util.List;
-
-public interface CollegeScoreCard {
-  /** General Use Functions */
-  boolean validateCollegeName(String collegeName);
-
-  List<UniversityData> getCollegesData(List<String> collegeNames);
-
-  /** Search Feature Functions */
-  List<UniversityData> getCollegeSuggestions(String currName);
-
-  /** Maps Feature Functions */
-  List<UniversityData> getMapData(List<UniversityData> collegeDatum);
-
-  /** DataVis Feature Functions */
-  List<UniversityData> getDataVisData(List<UniversityData> collegeDatum);
-}
