--- capstone/src/main/java/com/google/univiz/collegescorecarddata/MapData.java
@@ -1,8 +1,0 @@
-package com.google.univiz.collegescorecarddata;
-
-import java.util.List;
-
-public interface MapData {
-  /** Maps Feature Functions */
-  List<UniversityData> getMapData(List<UniversityData> collegeDatum);
-}
