--- capstone/src/main/java/com/google/univiz/collegescorecarddata/SearchData.java
@@ -1,8 +1,0 @@
-package com.google.univiz.collegescorecarddata;
-
-import java.util.List;
-
-public interface CollegeScoreCard {
-  /** Search Feature Functions */
-  List<UniversityData> getCollegeSuggestions(String currName);
-}
