--- capstone/src/main/java/com/google/univiz/collegescorecarddata/VisData.java
@@ -1,8 +1,0 @@
-package com.google.univiz.collegescorecarddata;
-
-import java.util.List;
-
-public interface CollegeScoreCard {
-  /** DataVis Feature Functions */
-  List<UniversityData> getDataVisData(List<UniversityData> collegeDatum);
-}
